// frontend/src/app/models/artikal.model.ts
export class Artikal{
    sifra: string;
    naziv: string;
    jedinica: string;
    stopa: number;
    kategorija: string;
    porijeklo: string;
    straniNaziv: string;
    barkod: number;
    proizvodjac: string;
    tarifa: string;
    eko: boolean;
    akcize: boolean;
    minZalihe: number;
    maxZalihe: number;
    opis: string;
    deklaracija: string;
    pib: string;
    slika: string;
}